Figure out a reasonable build process for a project of this scale.
Well, management decided that the thcrap GUI should be written in
Electron, so here I am, using this project as a change to finally write
something serious in JavaScript.

And even after just 10 lines, it became clear that I absolutely wanted
TypeScript together with an editor with first-class support for it if I
dont't want to go insane. So, after spending two days of sifting through
poor, outdated documentation, figuring out how you're actually supposed
to build this *without* resorting to some huge framework or build system
that I don't see the point in at this stage, and to debug it as
comfortably as C++ and C# code in Visual Studio, this is what I've ended
up with.

(Although we'll probably be better off in the future if we install
typescript locally via `npm install` after all, but I couldn't be
bothered to figure out how to make Visual Studio Code run *anything*
from node_modules…)